Move swagger to separate app.
