+++ apps/swagger/static/swagger-ui-standalone-preset.js
