+++ apps/swagger/static/swagger-ui.css
