+++ apps/swagger/static/swagger-ui.js
