+++ apps/swagger/templates/index.html
@@ -4,9 +4,9 @@
   <head>
     <meta charset="UTF-8">
     <title>Swagger UI</title>
-    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='swagger-ui.css') }}" >
-    <link rel="icon" type="image/png" href="{{ url_for('static', filename='favicon-32x32.png') }}" sizes="32x32" />
-    <link rel="icon" type="image/png" href="{{ url_for('static', filename='favicon-16x16.png') }}" sizes="16x16" />
+    <link rel="stylesheet" type="text/css" href="{{ url_for('swagger.static', filename='swagger-ui.css') }}" >
+    <link rel="icon" type="image/png" href="{{ url_for('swagger.static', filename='favicon-32x32.png') }}" sizes="32x32" />
+    <link rel="icon" type="image/png" href="{{ url_for('swagger.static', filename='favicon-16x16.png') }}" sizes="16x16" />
     <style>
       html
       {
@@ -33,13 +33,13 @@
   <body>
     <div id="swagger-ui"></div>
 
-    <script src="{{ url_for('static', filename='swagger-ui-bundle.js') }}"> </script>
-    <script src="{{ url_for('static', filename='swagger-ui-standalone-preset.js') }}"> </script>
+    <script src="{{ url_for('swagger.static', filename='swagger-ui-bundle.js') }}"> </script>
+    <script src="{{ url_for('swagger.static', filename='swagger-ui-standalone-preset.js') }}"> </script>
     <script>
     window.onload = function() {
       // Begin Swagger UI call region
       const ui = SwaggerUIBundle({
-        url: "{{ url_for('projects.spec_data') }}",
+        url: "{{ url_for('swagger.spec_data') }}",
         dom_id: '#swagger-ui',
         deepLinking: true,
         presets: [
