+++ settings.py
@@ -35,14 +35,13 @@ def celery_queue(name):
     """
     return "{}{}".format(os.environ.get('VIDEOSERVER_CELERY_PREFIX', ''), name)
 
-
 #: logging
 LOG_CONFIG_FILE = env('LOG_CONFIG_FILE', 'logging_config.yml')
 
 CORE_APPS = [
-    'apps',
-    'media',
-    'media.video',
+    'apps.swagger',
+    # 'media',
+    # 'media.video',
 ]
 
 #: Mongo host port
@@ -66,6 +65,3 @@ CODEC_SUPPORT = env('CODEC_SUPPORT', ['vp8', 'vp9', 'h264', 'aac', 'flac', 'ogg'
 
 #: media storage
 MEDIA_STORAGE = env('MEDIA_STORAGE', 'filesystem')
-
-#: log config file
-LOG_CONFIG_FILE = env('LOG_CONFIG_FILE', 'logging_config.yml')
